# assets/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Documentation API_Express",
      "version": "1.0.0",
      "license": {
        "name": "ISC"
      },
      "description": "API REST pour gérer des utilisateurs"
    },
    "basePath": "/api/v1",
    "host": "localhost:8080",
    "tags": [
      {
        "name": "members",
        "description": "Méthodes permettant de gérer les utilisateurs"
      }
    ],
    "paths": {
      "/api/v1/members/{id}": {
        "get": {
          "summary": "members/{id}",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {},
          "tags": ["members"]
        },
        "put": {
          "summary": "members/{id}",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {},
          "tags": ["members"]
        },
        "delete": {
          "summary": "members/{id}",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {},
          "tags": ["members"]
        }
      },
      "/api/v1/members": {
        "get": {
          "summary": "members",
          "responses": {},
          "tags": ["members"],
          "parameters": [
            {
              "name": "max",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer"
              }
            }
          ]
        },
        "post": {
          "summary": "members",
          "responses": {},
          "tags": ["members"],
          "parameters": [
            {
              "name": "name",
              "in": "body",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ]
        }
      }
    }
  }